added JSON try/catch and fix function for list.json
Added some basic error checking for the list.json file to make sure there is valid json in the file and prompts the user if it is not present. In addition to this, I have added a new option in the bash script (to-do.sh) - f, which rewrites the initial json to the file, which can also be used to completely delete the list.