Replacing init.sh with installer.sh
Installer.sh replaces the old init.sh, and provides additional functionality, including setting up custom file paths and properly validating those selected paths. 

The README.md has also be updated a bit, to provide more insight into the installer. It will probably be overhauled soon, to be a bit quicker to read, and more in depth documentation moved to a directory specifically for it.