Fixing error in to-do.sh and some other fixes
On the last commit, I had incorrectly changed how so things worked but did not properly test out the changes. It turns out, that I had not changed the path for the config file in the bash script, and the new one was not functioning. This commit removes the old file, properly sets up the new on and removes a test file that I made but did not remove.