Fixed 'f' option and problem in installer.sh
Unfortunately, I had pushed some other fixes without checking to see if the full thing worked. I have fixed the issue with the 'f' option writing invalid JSON and then the program prompting the user to rerun it. The list was missing some '' so after adding them, it worked as expected.